refactor(tui::ui::components::popups): move "ErrorPopup" to own file
