+++ tui/src/ui/components/popups/mod.rs
@@ -1,3 +1,5 @@
+#![allow(clippy::module_name_repetitions)]
+
 use crate::config::{BindingForEvent, StyleColorSymbol};
 /**
  * MIT License
@@ -32,6 +34,10 @@ use tuirealm::props::{
     TextModifiers, TextSpan,
 };
 use tuirealm::{AttrValue, Attribute, Component, Event, MockComponent, State, StateValue};
+
+mod error;
+
+pub use error::ErrorPopup;
 
 #[derive(MockComponent)]
 pub struct QuitPopup {
@@ -146,46 +152,6 @@ impl Component<Msg, NoUserEvent> for QuitPopup {
             Some(Msg::QuitPopupCloseOk)
         } else {
             Some(Msg::None)
-        }
-    }
-}
-
-#[derive(MockComponent)]
-pub struct ErrorPopup {
-    component: Paragraph,
-}
-
-impl ErrorPopup {
-    pub fn new<E: Into<anyhow::Error>>(msg: E) -> Self {
-        let msg = msg.into();
-        error!("Displaying error popup: {msg:?}");
-        // TODO: Consider changing to ":?" to output "Caused By" (and possibly backtrace) OR do a custom printing (copied from anyhow) once more than 4 lines can be displayed in height
-        let msg = format!("{msg:#}");
-        Self {
-            component: Paragraph::default()
-                .borders(
-                    Borders::default()
-                        .color(Color::Red)
-                        .modifiers(BorderType::Rounded),
-                )
-                .title(" Error ", Alignment::Center)
-                .foreground(Color::Red)
-                // .background(Color::Black)
-                .modifiers(TextModifiers::BOLD)
-                .alignment(Alignment::Center)
-                .text(&[TextSpan::from(msg)]/* &msg.lines().map(|v| TextSpan::from(v)).collect::<Vec<_>>() */),
-        }
-    }
-}
-
-impl Component<Msg, NoUserEvent> for ErrorPopup {
-    fn on(&mut self, ev: Event<NoUserEvent>) -> Option<Msg> {
-        match ev {
-            Event::Keyboard(KeyEvent {
-                code: Key::Enter | Key::Esc,
-                ..
-            }) => Some(Msg::ErrorPopupClose),
-            _ => None,
         }
     }
 }
